requirements.txt
flask>=3.0
python-dotenv>=1.0
resend>=0.7
